updating 4/4/21

+++ function.zsh
@@ -1,25 +1,34 @@
-venv() { python"$1" -m virtualenv venv --always-copy; }
+venv() { python"$1" -m virtualenv venv --always-copy }
 
-act() {
-if [[ -a "env.sh" ]]; then
-    source env.sh;
-fi
+run() {
+    if [[ -f ".env" ]]
+    then
+        source .env
+    fi
 
-source venv/bin/activate;
-code .;
+    if [[ -e "./venv/bin/activate" ]]
+    then
+        source venv/bin/activate
+    fi
+    code .
 }
 
 bye() {
-deactivate;
-cd ~/projects/;
-ls -l;
+    deactivate > /dev/null 2>&1
+    cd ~/projects/
+    ls -l
 }
 
 grr() {
-  git reset --hard "$1"/"$2";
-  git pull "$1" "$2";
+    if git rev-parse --git-dir > /dev/null 2>&1
+    then
+        git reset --hard "$1"/"$2"
+        git pull "$1" "$2"
+    else
+        echo "$PWD is not a git repository"
+    fi
 }
 
 magic8() {
-  python3 ~/projects/python/magic-8-ball/main.py;
+    python3 ~/projects/python/magic-8-ball/main.py
 }
